requirements.txt
scikit-image
opencv2
numpy
pandas
imutils